# docs/api/openapi.yml
openapi: 3.0.3
info:
  description: OpenTelemetry-Native Logs, Metrics and Traces in a single pane
  title: SigNoz
  version: ""
paths:
...
  /api/v1/dashboards/{id}/public:
    delete:
      deprecated: false
      description: This endpoint deletes the public sharing config and disables the
        public sharing of a dashboard
      operationId: DeletePublicDashboard
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          content:
            application/json:
              schema:
                type: string
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: Delete public dashboard
      tags:
      - dashboard
    get:
      deprecated: false
      description: This endpoint returns public sharing config for a dashboard
      operationId: GetPublicDashboard
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/DashboardtypesGettablePublicDasbhboard'
                  status:
                    type: string
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: Get public dashboard
      tags:
      - dashboard
    post:
      deprecated: false
      description: This endpoint creates public sharing config and enables public
        sharing of the dashboard
      operationId: CreatePublicDashboard
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardtypesPostablePublicDashboard'
      responses:
        "201":
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/TypesIdentifiable'
                  status:
                    type: string
                type: object
          description: Created
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: Create public dashboard
      tags:
      - dashboard
    put:
      deprecated: false
      description: This endpoint updates the public sharing config for a dashboard
      operationId: UpdatePublicDashboard
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardtypesUpdatablePublicDashboard'
      responses:
        "204":
          content:
            application/json:
              schema:
                type: string
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: Update public dashboard
      tags:
      - dashboard
...
  /api/v1/public/dashboards/{id}:
    get:
      deprecated: false
      description: This endpoint returns the sanitized dashboard data for public access
      operationId: GetPublicDashboardData
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/DashboardtypesGettablePublicDashboardData'
                  status:
                    type: string
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - anonymous:
        - public-dashboard:read
      summary: Get public dashboard data
      tags:
      - dashboard
  /api/v1/public/dashboards/{id}/widgets/{idx}/query_range:
    get:
      deprecated: false
      description: This endpoint return query range results for a widget of public
        dashboard
      operationId: GetPublicDashboardWidgetQueryRange
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      - in: path
        name: idx
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/Querybuildertypesv5QueryRangeResponse'
                  status:
                    type: string
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - anonymous:
        - public-dashboard:read
      summary: Get query range result
      tags:
      - dashboard
...
components:
  schemas:
...
    DashboardtypesDashboard:
      properties:
        createdAt:
          format: date-time
          type: string
        createdBy:
          type: string
        data:
          $ref: '#/components/schemas/DashboardtypesStorableDashboardData'
        id:
          type: string
        locked:
          type: boolean
        org_id:
          type: string
        updatedAt:
          format: date-time
          type: string
        updatedBy:
          type: string
      type: object
    DashboardtypesGettablePublicDasbhboard:
      properties:
        defaultTimeRange:
          type: string
        publicPath:
          type: string
        timeRangeEnabled:
          type: boolean
      type: object
    DashboardtypesGettablePublicDashboardData:
      properties:
        dashboard:
          $ref: '#/components/schemas/DashboardtypesDashboard'
        publicDashboard:
          $ref: '#/components/schemas/DashboardtypesGettablePublicDasbhboard'
      type: object
    DashboardtypesPostablePublicDashboard:
      properties:
        defaultTimeRange:
          type: string
        timeRangeEnabled:
          type: boolean
      type: object
    DashboardtypesStorableDashboardData:
      additionalProperties: {}
      type: object
    DashboardtypesUpdatablePublicDashboard:
      properties:
        defaultTimeRange:
          type: string
        timeRangeEnabled:
          type: boolean
      type: object
    ErrorsJSON:
      properties:
        code:
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorsResponseerroradditional'
          type: array
        message:
          type: string
        url:
          type: string
      type: object
    ErrorsResponseerroradditional:
      properties:
        message:
          type: string
      type: object
...
    Querybuildertypesv5ExecStats:
      properties:
        bytesScanned:
          minimum: 0
          type: integer
        durationMs:
          minimum: 0
          type: integer
        rowsScanned:
          minimum: 0
          type: integer
        stepIntervals:
          additionalProperties:
            minimum: 0
            type: integer
          type: object
      type: object
...
    Querybuildertypesv5QueryData:
      properties:
        results:
          items: {}
          nullable: true
          type: array
      type: object
    Querybuildertypesv5QueryRangeResponse:
      properties:
        data:
          $ref: '#/components/schemas/Querybuildertypesv5QueryData'
        meta:
          $ref: '#/components/schemas/Querybuildertypesv5ExecStats'
        type:
          type: string
        warning:
          $ref: '#/components/schemas/Querybuildertypesv5QueryWarnData'
      type: object
    Querybuildertypesv5QueryWarnData:
      properties:
        message:
          type: string
        url:
          type: string
        warnings:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5QueryWarnDataAdditional'
          type: array
      type: object
    Querybuildertypesv5QueryWarnDataAdditional:
      properties:
        message:
          type: string
      type: object
    RenderErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorsJSON'
        status:
          type: string
      type: object
...
    TypesIdentifiable:
      properties:
        id:
          type: string
      type: object
...
  securitySchemes:
    api_key:
      description: API Keys
      in: header
      name: SigNoz-Api-Key
      type: apiKey
    tokenizer:
      bearerFormat: Tokenizer
      description: Tokens generated by the tokenizer
      scheme: bearer
      type: http